// frontend/src/app/core/api/generated/model/page-response-inventory-setting-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InventorySettingResponse } from './inventory-setting-response';


export interface PageResponseInventorySettingResponse { 
    content?: Array<InventorySettingResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    first?: boolean;
    last?: boolean;
}